+++ packages/data-components/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.0.0] - 2026-01-26
+
+- Update peer dependencies to React 18
+- Update bundling process to `@macrostrat/web-components-bundler`
+- Make CSS imports optional; users must now import
+  `@macrostrat/data-components/style.css` or
+  `@macrostrat/data-components/dist/data-components.css` for styles
+
 ## [0.3.0] - 2026-01-20
 
 - Replace `node-fetch` with `cross-fetch` in `PrevalentTaxa` component